Fix: Add .NET SDK validation and fix SQLite type mapping
After @v0l pointed out that the code wasn't tested with dotnet, I've:

1. Verified the entire project builds and all 23 tests pass with .NET SDK
2. Fixed ID retrieval in InsertPartials/InsertPartialsAsync with proper type conversion
3. Updated SQLite example to use compatible types (double for Price, long for Stock)
   to match SQLite's type system

All changes have been validated by actually building and running the tests.

+++ examples/SQLiteExample/Program.cs
@@ -114,13 +114,13 @@ public class Product
     public int Id { get; set; }
 
     public Partial<string> Name { get; set; }
-    public Partial<decimal> Price { get; set; }
-    public Partial<int> Stock { get; set; }
+    public Partial<double> Price { get; set; }
+    public Partial<long> Stock { get; set; }
     public Partial<string> Description { get; set; }
 
     // Helper properties for reading values in the example
     public string? NameValue => Name.IsSet ? Name.Value : null;
-    public decimal? PriceValue => Price.IsSet ? Price.Value : null;
-    public int? StockValue => Stock.IsSet ? Stock.Value : null;
+    public double? PriceValue => Price.IsSet ? Price.Value : null;
+    public long? StockValue => Stock.IsSet ? Stock.Value : null;
     public string? DescriptionValue => Description.IsSet ? Description.Value : null;
 }
